feat: DaoGeneration feature cleanup
