+++ frontend/src/lib/features/dao-generator/components/index.ts
@@ -1,0 +1,4 @@
+export { default as RecapCard } from './atoms/RecapCard.svelte';
+export { default as RecapElement } from './atoms/RecapElement.svelte';
+export { default as StepButtons } from './atoms/StepButtons.svelte';
+export { default as GeneratorNav } from './navigation/GeneratorNav.svelte';
