+++ frontend/src/lib/features/dao-generator/steps/index.ts
@@ -1,0 +1,5 @@
+export { default as GeneralData } from './1-general-data/GeneralData.svelte';
+export { default as Description } from './2-description/Description.svelte';
+export { default as Socials } from './3-socials/Socials.svelte';
+export { default as Tokenomics } from './4-tokenomics/Tokenomics.svelte';
+export { default as ReviewAndDeploy } from './5-review-and-deploy/ReviewAndDeploy.svelte';
